additional folder refactoring
