+++ app/slack-messages/index.ts
@@ -1,6 +1,5 @@
 export * from './delete-initiative';
 export * from './edit-initiative';
-export * from './id-helper';
 export * from './initiative-details';
 export * from './initiative-list';
 export * from './new-member';
