+++ app/slack-messages/shared-messages/index.ts
@@ -1,0 +1,2 @@
+export * from './members';
+export * from './initiatives';
