+++ app/slash-commands/add-initiative.ts
@@ -1,9 +1,9 @@
 import { DynamoDB } from 'aws-sdk';
 import { apiWrapper, ApiSignature } from '@manwaring/lambda-wrapper';
-import { CreateInitiativeRequest, InitiativeResponse, InitiativeRecord, INITIATIVE_TYPE } from '../initiative';
+import { CreateInitiativeRequest, InitiativeResponse, InitiativeRecord, INITIATIVE_TYPE } from '../common/initiative';
 import { getUserProfile } from '../slack-api';
 import { DetailResponse } from '../slack-messages';
-import { MemberResponse, MEMBER_TYPE, getTeamIdentifier } from '../member';
+import { MemberResponse, MEMBER_TYPE, getTeamIdentifier } from '../common/member';
 import { SlashCommandBody } from 'slack';
 
 const initiatives = new DynamoDB.DocumentClient({ region: process.env.REGION });
