+++ app/status-updates/request-update.ts
@@ -1,8 +1,8 @@
 import { DynamoDB } from 'aws-sdk';
 import { snsWrapper, SnsSignature } from '@manwaring/lambda-wrapper';
-import { InitiativeRecord, InitiativeResponse, INITIATIVE_TYPE } from '../initiative';
-import { MemberResponse, MEMBER_TYPE } from '../member';
-import { send } from '../slack-api';
+import { InitiativeRecord, InitiativeResponse, INITIATIVE_TYPE } from '../common/initiative';
+import { MemberResponse, MEMBER_TYPE } from '../common/member';
+import { sendMessage } from '../slack-api';
 import { StatusUpdateRequest } from '../slack-messages';
 
 const initiatives = new DynamoDB.DocumentClient({ region: process.env.REGION });
@@ -14,7 +14,7 @@ export const handler = snsWrapper(async ({ message, success, error }: SnsSignatu
     await Promise.all(
       champions
         .map(champion => new StatusUpdateRequest(initiative, champion))
-        .map(request => send(request, initiative.team.id))
+        .map(request => sendMessage(request, initiative.team.id))
     );
     success();
   } catch (err) {
